+++ app/api_v1/__init__.py
@@ -4,5 +4,9 @@ api = Blueprint('api_v1', __name__)
 from . import authentication, errors, utils
 from .errors import *
 from .index import *
+from .rate_limit import *
+from .etag import *
+from .utils import *
 from app.api_v1.resources import *
 from app.api_v1.fhir import *
+from .errors_test import *
