+++ app/api_v1/errors_test/exceptions.py
@@ -1,0 +1,4 @@
+from ..errors import AuthenticationError
+
+class TokenExpiredError(AuthenticationError):
+    pass
