+++ app/api_v1/etag.py
@@ -1,0 +1,34 @@
+import functools
+import hashlib
+
+from flask import request, make_response
+
+from app.api_v1.errors import precondition_failed, not_modified
+
+
+def etag(f):
+    """This decorator adds an ETag header to the response."""
+
+    @functools.wraps(f)
+    def wrapped(*args, **kwargs):
+        # only for HEAD and GET requests
+        assert request.method in ['HEAD', 'GET'], \
+            '@etag is only supported for GET requests'
+        rv = f(*args, **kwargs)
+        rv = make_response(rv)
+        etag = '"' + hashlib.md5(rv.get_data()).hexdigest() + '"'
+        rv.headers['Cache-Control'] = 'max-age=86400'
+        rv.headers['ETag'] = etag
+        if_match = request.headers.get('If-Match')
+        if_none_match = request.headers.get('If-None-Match')
+        if if_match:
+            etag_list = [tag.strip() for tag in if_match.split(',')]
+            if etag not in etag_list and '*' not in etag_list:
+                rv = precondition_failed()
+        elif if_none_match:
+            etag_list = [tag.strip() for tag in if_none_match.split(',')]
+            if etag in etag_list or '*' in etag_list:
+                rv = not_modified()
+        return rv
+
+    return wrapped
