+++ app/api_v1/fhir/resources/CodeSystem.py
@@ -1,9 +1,7 @@
-from flask import url_for
-
 from app.api_v1.authentication import token_auth
 from app.api_v1.errors import *
 from app.api_v1.rate_limit import rate_limit
-from app.api_v1.utils import etag
+from app.api_v1.etag import etag
 from app.models.fhir.codesets import CodeSystem
