+++ app/api_v1/fhir/resources/Patient.py
@@ -1,10 +1,12 @@
-from flask import url_for, request
+from flask import request
 
 from app.api_v1.authentication import token_auth
 from app.api_v1.errors import *
 from app.api_v1.rate_limit import rate_limit
-from app.api_v1.utils import etag
-from app.api_v1.fhir.fhir_utils import create_bundle, parse_fhir_search, enforce_fhir_mimetype_charset
+from app.api_v1 import etag
+from app.api_v1.fhir.utils.bundle_utils import create_bundle
+from app.api_v1.fhir.utils.search_utils import parse_fhir_search
+from app.api_v1.fhir.utils.request_utils import enforce_fhir_mimetype_charset
 from app.models.fhir.patient import Patient
