+++ app/api_v1/fhir/resources/__init__.py
@@ -1,0 +1,3 @@
+from .CodeSystem import *
+from .Patient import *
+from .ValueSet import *
