+++ app/api_v1/resources/__init__.py
@@ -1,1 +1,1 @@
-from . import User
+from .User import *
